+++ scripts/legacyMacros.js
@@ -569,6 +569,7 @@ export {squeeze} from './macros/2014/actions/squeeze.js';
 export {stabilize} from './macros/2014/actions/stabilize.js';
 export {underwater, underwaterEffect} from './macros/2014/actions/underwater.js';
 export {genericActions} from './macros/2014/actions/genericActions.js';
+export {healingSurge} from './macros/2014/actions/healingSurge.js';
 // Feats
 export {aberrantDragonmark} from './macros/2014/feats/aberrantDragonmark.js';
 export {bulwark} from './macros/2014/feats/bulwark.js';