+++ scripts/macros/2014/actions/genericActions.js
@@ -1,7 +1,6 @@
 import {genericActions as modernGenericActions} from '../../2024/actions/genericActions.js';
 export let genericActions = {
-    name: 'Generic Actions',
+    ...modernGenericActions,
     version: '1.3.115',
     rules: 'legacy',
-    ...modernGenericActions
 };